ejercicio 3 reparado

+++ practica2/ejercicio3.cpp
@@ -5,7 +5,7 @@ cout<<"Introduzca un número para conocer su factorial"<<endl;
 
 int num;
 int f;
-int resultado = 1;
+double resultado = 1;
 
 cin>>num;
